refactor: move client and prompt typevars out of types
