+++ python/mirascope/llm/calls/base_call.py
@@ -6,8 +6,9 @@ from dataclasses import dataclass
 from typing import Generic
 
 from ..models import LLM
+from ..prompts import PromptT
 from ..tools import ToolDef
-from ..types import Jsonable, P, PromptT
+from ..types import Jsonable, P
 
 
 @dataclass
