+++ python/mirascope/llm/calls/base_structured_context_call.py
@@ -6,8 +6,9 @@ from dataclasses import dataclass
 from typing import Generic
 
 from ..models import LLM
+from ..prompts import PromptT
 from ..tools import ContextToolDef
-from ..types import DepsT, FormatT, Jsonable, P, PromptT
+from ..types import DepsT, FormatT, Jsonable, P
 
 
 @dataclass
