+++ python/mirascope/llm/clients/__init__.py
@@ -1,7 +1,7 @@
 """Client interfaces for LLM providers."""
 
 from .anthropic import AnthropicClient, AnthropicMessage, AnthropicParams
-from .base import BaseClient, BaseParams
+from .base import LLMT, BaseClient, BaseParams, ClientT, ParamsT, ProviderMessageT
 from .google import GoogleClient, GoogleMessage, GoogleParams
 from .openai import OpenAIClient, OpenAIMessage, OpenAIParams
 from .register import (
@@ -14,6 +14,7 @@ from .register import (
 __all__ = [
     "ANTHROPIC_REGISTERED_LLMS",
     "GOOGLE_REGISTERED_LLMS",
+    "LLMT",
     "OPENAI_REGISTERED_LLMS",
     "REGISTERED_LLMS",
     "AnthropicClient",
@@ -21,10 +22,13 @@ __all__ = [
     "AnthropicParams",
     "BaseClient",
     "BaseParams",
+    "ClientT",
     "GoogleClient",
     "GoogleMessage",
     "GoogleParams",
     "OpenAIClient",
     "OpenAIMessage",
     "OpenAIParams",
+    "ParamsT",
+    "ProviderMessageT",
 ]
