+++ python/mirascope/llm/prompts/__init__.py
@@ -13,6 +13,7 @@ from .decorator import (
     ContextPrompt,
     MessagesPrompt,
     Prompt,
+    PromptT,
     prompt,
 )
 
@@ -25,5 +26,6 @@ __all__ = [
     "ContextPrompt",
     "MessagesPrompt",
     "Prompt",
+    "PromptT",
     "prompt",
 ]
